requirements.txt
fastapi==0.117.1
uvicorn==0.30.6
langgraph==0.2.16
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-ollama==0.1.3
langchain-openai==0.1.23
chromadb==0.5.5
python-docx==1.1.2
pydantic==2.8.2
python-dotenv==1.0.1
mcp==1.0.0
